+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/nationalarchives/tna-frontend/compare/v0.2.13...HEAD)
 
 ### Added
+
+- Items in a files list can be made entirely clickable using the `fullAreaClick` property
+
 ### Changed
 
 - Updated the default `aria-label` on breadcrumbs components from "Breadcrumb" to "Breadcrumbs"